+++ packages/terra-consumer-nav/src/components/user-profile/UserProfile.jsx
@@ -24,17 +24,13 @@ const propTypes = {
    */
   id: PropTypes.string,
   /**
-   * The path signout button would redirect to.
+   * The path signing out would redirect to. This link, when user data is available, will appear in the popup/modal, otherwise will display as the profile link itself.
    */
   signoutUrl: PropTypes.string.isRequired,
   /**
-   * The path to the login page.
+   * The path to the login page; this only gets displayed on the profile if no user information is provided along side this link
    */
   signinUrl: PropTypes.string,
-  /**
-   * Determiniate of whether profile should render as signin link or profile button.
-   */
-  isSignIn: PropTypes.bool,
   /**
    * The content of the each profile items.
    */
@@ -50,33 +46,47 @@ const propTypes = {
 };
 
 const defaultProps = {
-  avatar: <IconPerson />,
   profileLinks: [],
-  isSignIn: false,
   id: 'terra-conumser-nav-profile-button',
 };
 
 const UserProfile = ({
-  userName, avatar, id, signoutUrl, signinUrl, isSignIn, profileLinks, handleClick, intl, ...customProps
+  userName, avatar, id, signoutUrl, signinUrl, profileLinks, handleClick, intl, ...customProps
 }) => {
   let profileContent;
-  if (isSignIn) {
-    profileContent = (
-      <button className={cx('popup-button')} href={signinUrl}>
-        <Arrange
-          fitStart={<div className={cx('avatar')}>{avatar}</div>}
-          fill={<FormattedMessage id="Terra.Consumer.UserProfile.signin" />}
-          align="center"
-        />
-      </button>
-    );
+  const avatarIcon = <div className={cx('avatar')}>{avatar || <IconPerson />}</div>;
+
+  const singleProfileLink = (url, translationId) => (
+    <a href={url} className={cx('profile-text')}>
+      <Arrange
+        className={cx('popup-button')}
+        fitStart={avatarIcon}
+        fill={<FormattedMessage id={translationId} />}
+        align="center"
+      />
+    </a>
+  );
+
+  /*
+   * Logic tree:
+   *  If no user infomation is available, specifically the avatar and username, then one of two things will happen.
+   *   1: a signIn link was provided and therefore the sign IN option will be presented.
+   *   2: no signIn link was provided and as such the sign OUT link will be presented.
+   * Otherwise, the profile will display as normal with onclick opening the popup/modal depending on screensize,
+   *   desktop/modal respectively.
+   */
+  if (!(avatar || userName)) {
+    profileContent = signinUrl ?
+      singleProfileLink(signinUrl, 'Terra.Consumer.UserProfile.signin')
+    :
+      singleProfileLink(signoutUrl, 'Terra.Consumer.UserProfile.signout');
   } else {
     const content = (
       <div>
         <ProfileLinks linkItems={profileLinks} />
-        <button className={cx('link', 'signout-border')} href={signoutUrl}>
+        <a className={cx('link', 'signout-border')} href={signoutUrl}>
           <FormattedMessage id="Terra.Consumer.UserProfile.signout" />
-        </button>
+        </a>
       </div>
     );
 
@@ -85,7 +95,7 @@ const UserProfile = ({
     profileContent = (
       <button className={cx('popup-button')} onClick={() => handleClick({ title, content })}>
         <Arrange
-          fitStart={<div className={cx('avatar')}>{avatar}</div>}
+          fitStart={avatarIcon}
           fill={<span>{userName}</span>}
           fitEnd={<IconEllipses className={cx('icon')} id={id} />}
           align="center"