+++ packages/terra-consumer-nav/tests/nightwatch/NavTestRoutes.jsx
@@ -7,8 +7,9 @@ import NavTests from './NavTests';
 // Test Cases
 import BurgerButtonTest from './BurgerButtonTest';
 import HelpButtonTest from './HelpButtonTest';
-import SimpleNav from './SimpleNav';
 import NavLogoTest from './NavLogoTest';
+import ProfileButtonTest from './ProfileButtonTest';
+import SimpleNav from './SimpleNav';
 import SimpleNavWithTextLogo from './SimpleNavWithTextLogo';
 
 const routes = (
@@ -16,9 +17,10 @@ const routes = (
     <Route path="/tests/nav-tests" component={NavTests} />
     <Route path="/tests/nav-tests/burger-button" component={BurgerButtonTest} />
     <Route path="/tests/nav-tests/help-button" component={HelpButtonTest} />
+    <Route path="/tests/nav-tests/logo" component={NavLogoTest} />
+    <Route path="/tests/nav-tests/profile-button" component={ProfileButtonTest} />
     <Route path="/tests/nav-tests/simple-nav" component={SimpleNav} />
     <Route path="/tests/nav-tests/simple-nav-logo-text" component={SimpleNavWithTextLogo} />
-    <Route path="/tests/nav-tests/logo" component={NavLogoTest} />
   </div>
 );